rootdialog fix
