+++ SmartKioskBot/SmartKioskBot.csproj
@@ -212,8 +212,9 @@
     <Compile Include="Dialogs\ProductDetails.cs" />
     <Compile Include="Dialogs\QnA\QnADialog.cs" />
     <Compile Include="Dialogs\QnA\QnAMakerResult.cs" />
-    <Compile Include="Dialogs\RecommendationDialog.cs" />
     <Compile Include="Dialogs\RootDialog.cs" />
+    <Compile Include="Dialogs\RecommendationDialog.cs" />
+    <Compile Include="Dialogs\LuisProcessingDialog.cs" />
     <Compile Include="Dialogs\StockDialog.cs" />
     <Compile Include="Dialogs\WishListDialog.cs" />
     <Compile Include="Global.asax.cs">
